+++ tests/test_cross_section.py
@@ -349,6 +349,16 @@ def test_get_wetted_surface(cs):
     assert isinstance(pol3, geometry.MultiPolygon)
     assert len(pol3.geoms) == 1
 
+def test_get_wetted_surface_sz_perimeter(cs):
+    """Test get_wetted_surface with `perimeter=True`"""
+    line1 = cs.get_wetted_surface_sz(h=93.0, perimeter=True)
+    line2 = cs.get_wetted_surface_sz(h=94.9, perimeter=True)
+
+    assert isinstance(line1, geometry.MultiLineString)
+    assert line1.has_z == False
+    assert isinstance(line2, geometry.MultiLineString)
+    assert line2.has_z == False
+    assert line2.length > line1.length
 
 def test_detect_wl(cs, img):
     h = cs.detect_water_level(img, bank="far", length=1.0)